Moved related classes to one package
