+++ collect_app/src/main/java/org/odk/collect/android/mainmenu/MainMenuActivity.kt
@@ -19,7 +19,6 @@ import org.odk.collect.android.activities.DeleteSavedFormActivity
 import org.odk.collect.android.activities.FirstLaunchActivity
 import org.odk.collect.android.activities.FormDownloadListActivity
 import org.odk.collect.android.activities.InstanceChooserList
-import org.odk.collect.android.activities.InstanceUploaderListActivity
 import org.odk.collect.android.activities.WebViewActivity
 import org.odk.collect.android.application.MapboxClassInstanceCreator.createMapBoxInitializationFragment
 import org.odk.collect.android.application.MapboxClassInstanceCreator.isMapboxAvailable
@@ -29,6 +28,7 @@ import org.odk.collect.android.formmanagement.FormFillingIntentFactory
 import org.odk.collect.android.injection.DaggerUtils
 import org.odk.collect.android.projects.ProjectIconView
 import org.odk.collect.android.projects.ProjectSettingsDialog
+import org.odk.collect.android.readytosend.InstanceUploaderListActivity
 import org.odk.collect.android.utilities.ApplicationConstants
 import org.odk.collect.android.utilities.ThemeUtils
 import org.odk.collect.androidshared.ui.DialogFragmentUtils.showIfNotShowing
