+++ collect_app/src/main/java/org/odk/collect/android/notifications/builders/FormsSubmissionNotificationBuilder.kt
@@ -6,9 +6,9 @@ import android.app.PendingIntent
 import android.content.Intent
 import androidx.core.app.NotificationCompat
 import org.odk.collect.android.R
-import org.odk.collect.android.activities.InstanceUploaderListActivity
 import org.odk.collect.android.mainmenu.MainMenuActivity
 import org.odk.collect.android.notifications.NotificationManagerNotifier
+import org.odk.collect.android.readytosend.InstanceUploaderListActivity
 import org.odk.collect.android.upload.FormUploadException
 import org.odk.collect.android.utilities.ApplicationConstants.RequestCodes
 import org.odk.collect.android.utilities.FormsUploadResultInterpreter
